+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 Current
 -------
 
+- Add authorizations parsing to namespace (:issue:`403`)
 - Add vendor extensions support (:issue:`97`)
 - :class:`~reqparse.RequestParser` arguments now support the ``split`` action
 - Ensure default boolean value as `False` works with :class:`~reqparse.RequestParser` (:issue:`199`)